More no-op code cleanup (#1222)
* Move handlers & routes for results receiver

into the receiver package for consistency.

* Rejig AppEngineAPI interface and its impl

1. Un-export shared.AppEngineAPIImpl and move its documentation to the
   interface.
2. Correctly use pointer receiver type when we need to modify
   shared.appEngineAPIImpl or receiver.apiImpl.

* Delegate GetUploader to Datastore

Also call GetUploader in receiver.API instead of crafting another
Datastore query.

* Move azure.HandleCheckRunEvent into api.go

along with its test. One benefit is that api/azure/webhook_test.go now
no longer needs to be in a separate package to avoid cyclic dependency
and a bunch of regexps no longer need to be exported.
