+++ shared/appengine.go
@@ -18,7 +18,6 @@ import (
 	"github.com/google/go-github/github"
 	"golang.org/x/oauth2"
 	"google.golang.org/appengine"
-	"google.golang.org/appengine/datastore"
 	"google.golang.org/appengine/urlfetch"
 	"google.golang.org/appengine/user"
 )
@@ -27,70 +26,79 @@ import (
 type AppEngineAPI interface {
 	Context() context.Context
 
+	// GitHub OAuth client
+	GetGitHubClient() (*github.Client, error)
+	// urlfetch cilents
 	GetHTTPClient() *http.Client
 	GetSlowHTTPClient(time.Duration) (*http.Client, context.CancelFunc)
-	GetGitHubClient() (*github.Client, error)
 
-	// The three methods below are exported for webapp.admin_handler.
-	IsLoggedIn() bool
+	// AppEngine User API
 	IsAdmin() bool
+	IsLoggedIn() bool
 	LoginURL(redirect string) (string, error)
-
-	IsFeatureEnabled(featureName string) bool
-	GetUploader(uploader string) (Uploader, error)
 
 	// GetVersion returns the version name for the current environment.
 	GetVersion() string
-	// GetHostname returns the canonical hostname for the current appengine project,
-	// i.e. staging.wpt.fyi or wpt.fyi
+	// GetHostname returns the canonical hostname for the current AppEngine project,
+	// i.e. staging.wpt.fyi or wpt.fyi.
 	GetHostname() string
-	// GetVersionedHostname returns the canonical hostname for the current version,
-	// i.e. version-dot-wptdashboard{,-staging}.appspot.com
+	// GetVersionedHostname returns the AppEngine hostname for the current version,
+	// i.e. version-dot-wptdashboard{,-staging}.appspot.com.
 	GetVersionedHostname() string
+	// GetServiceHostname returns the AppEngine hostname for the given service (module) of the
+	// current version,
+	// i.e. version-dot-service-dot-wptdashboard{,-staging}.appspot.com.
+	// Note that if the current version does not have the specified service, AppEngine routing
+	// will fall back to its default version.
 	GetServiceHostname(service string) string
+
+	// GetResultsURL returns a URL for the wpt.fyi results page for the test runs loaded for the
+	// given filter.
 	GetResultsURL(filter TestRunFilter) *url.URL
+	// GetRunsURL returns a URL for the wpt.fyi results page for the test runs loaded for the
+	// given filter.
 	GetRunsURL(filter TestRunFilter) *url.URL
+	// GetResultsUploadURL returns a URL for uploading results (i.e. results receiver).
 	GetResultsUploadURL() *url.URL
+
+	// Simple wrappers that delegate to Datastore
+	IsFeatureEnabled(featureName string) bool
+	GetUploader(uploader string) (Uploader, error)
 }
 
 // NewAppEngineAPI returns an AppEngineAPI for the given context.
-// Note that the context should be created using NewAppEngineContext
+// Note that the context should be created using NewAppEngineContext.
 func NewAppEngineAPI(ctx context.Context) AppEngineAPI {
-	return AppEngineAPIImpl{
+	return &appEngineAPIImpl{
 		ctx: ctx,
 	}
 }
 
-// AppEngineAPIImpl implements the AppEngineAPI interface.
-type AppEngineAPIImpl struct {
+// appEngineAPIImpl implements the AppEngineAPI interface.
+type appEngineAPIImpl struct {
 	ctx context.Context
-	// Cached client objects.
+	// Cached client objects
 	httpClient   *http.Client
 	githubClient *github.Client
 }
 
-// Context returns the context.Context for the API impl.
-func (a AppEngineAPIImpl) Context() context.Context {
+func (a appEngineAPIImpl) Context() context.Context {
 	return a.ctx
 }
 
-// GetHTTPClient returns an HTTP client in the current context.
-func (a AppEngineAPIImpl) GetHTTPClient() *http.Client {
+func (a *appEngineAPIImpl) GetHTTPClient() *http.Client {
 	if a.httpClient == nil {
 		a.httpClient = urlfetch.Client(a.ctx)
 	}
 	return a.httpClient
 }
 
-// GetSlowHTTPClient returns an HTTP client without timeout for the current
-// context.
-func (a AppEngineAPIImpl) GetSlowHTTPClient(timeout time.Duration) (*http.Client, context.CancelFunc) {
+func (a appEngineAPIImpl) GetSlowHTTPClient(timeout time.Duration) (*http.Client, context.CancelFunc) {
 	slowCtx, cancel := context.WithTimeout(a.ctx, timeout)
 	return urlfetch.Client(slowCtx), cancel
 }
 
-// GetGitHubClient returns a github client using the stored API token.
-func (a AppEngineAPIImpl) GetGitHubClient() (*github.Client, error) {
+func (a *appEngineAPIImpl) GetGitHubClient() (*github.Client, error) {
 	if a.githubClient == nil {
 		ds := NewAppEngineDatastore(a.ctx, false)
 		secret, err := GetSecret(ds, "github-api-token")
@@ -106,40 +114,29 @@ func (a AppEngineAPIImpl) GetGitHubClient() (*github.Client, error) {
 	return a.githubClient, nil
 }
 
-// IsLoggedIn returns true if a user is logged in for the current context.
-func (a AppEngineAPIImpl) IsLoggedIn() bool {
+func (a appEngineAPIImpl) IsLoggedIn() bool {
 	return user.Current(a.ctx) != nil
 }
 
-// LoginURL returns the URL for a user to log in.
-func (a AppEngineAPIImpl) LoginURL(redirect string) (string, error) {
+func (a appEngineAPIImpl) LoginURL(redirect string) (string, error) {
 	return user.LoginURL(a.ctx, redirect)
 }
 
-// IsAdmin returns true if the context's user is an Admin user.
-func (a AppEngineAPIImpl) IsAdmin() bool {
+func (a appEngineAPIImpl) IsAdmin() bool {
 	return user.IsAdmin(a.ctx)
 }
 
-// IsFeatureEnabled returns true if the given feature name is stored as a Flag
-// and its Enabled property is true.
-func (a AppEngineAPIImpl) IsFeatureEnabled(featureName string) bool {
-	// TODO(lukebjerring): Migrate other callers of this signature to AppEngineAPI
+func (a appEngineAPIImpl) IsFeatureEnabled(featureName string) bool {
 	ds := NewAppEngineDatastore(a.ctx, false)
 	return IsFeatureEnabled(ds, featureName)
 }
 
-// GetUploader returns the uploader with the given name.
-func (a AppEngineAPIImpl) GetUploader(uploader string) (Uploader, error) {
-	result := Uploader{}
-	key := datastore.NewKey(a.ctx, "Uploader", uploader, 0, nil)
-	err := datastore.Get(a.ctx, key, &result)
-	return result, err
+func (a appEngineAPIImpl) GetUploader(uploader string) (Uploader, error) {
+	ds := NewAppEngineDatastore(a.ctx, false)
+	return GetUploader(ds, uploader)
 }
 
-// GetHostname returns the canonical hostname for the current appengine project,
-// i.e. staging.wpt.fyi or wpt.fyi
-func (a AppEngineAPIImpl) GetHostname() string {
+func (a appEngineAPIImpl) GetHostname() string {
 	hostname := appengine.DefaultVersionHostname(a.ctx)
 	if hostname == "wptdashboard.appspot.com" {
 		return "wpt.fyi"
@@ -149,8 +146,7 @@ func (a AppEngineAPIImpl) GetHostname() string {
 	return hostname
 }
 
-// GetVersion returns the version name for the current environment.
-func (a AppEngineAPIImpl) GetVersion() string {
+func (a appEngineAPIImpl) GetVersion() string {
 	version := strings.Split(appengine.VersionID(a.ctx), ".")[0]
 	if appengine.IsDevAppServer() {
 		out, err := exec.Command("/usr/bin/git", "rev-parse", "--abbrev-ref", "HEAD").Output()
@@ -163,16 +159,12 @@ func (a AppEngineAPIImpl) GetVersion() string {
 	return version
 }
 
-// GetVersionedHostname returns the canonical hostname for the current version,
-// i.e. version-dot-wptdashboard{,-staging}.appspot.com
-func (a AppEngineAPIImpl) GetVersionedHostname() string {
+func (a appEngineAPIImpl) GetVersionedHostname() string {
 	hostname := appengine.DefaultVersionHostname(a.ctx)
 	return fmt.Sprintf("%s-dot-%s", a.GetVersion(), hostname)
 }
 
-// GetServiceHostname returns a versioned canonical hostname for the given service (module).
-// If the service does not have the current version, AppEngine routing will fall back to its default version.
-func (a AppEngineAPIImpl) GetServiceHostname(service string) string {
+func (a appEngineAPIImpl) GetServiceHostname(service string) string {
 	// version and instance (last 2 params) left blank means that the version of the current
 	// instance will be used. This is desirable for branches that push multiple services, and
 	// means that we should keep service production versions in sync.
@@ -185,26 +177,19 @@ func (a AppEngineAPIImpl) GetServiceHostname(service string) string {
 	return fmt.Sprintf("%s-dot-%s-dot-%s", a.GetVersion(), service, hostname)
 }
 
-// GetResultsURL returns a url for the wpt.fyi results page for the test runs
-// loaded for the given filter.
-func (a AppEngineAPIImpl) GetResultsURL(filter TestRunFilter) *url.URL {
+func (a appEngineAPIImpl) GetResultsURL(filter TestRunFilter) *url.URL {
 	return getURL(a.GetHostname(), "/results/", filter)
 }
 
-// GetRunsURL returns a url for the wpt.fyi results page for the test runs
-// loaded for the given filter.
-func (a AppEngineAPIImpl) GetRunsURL(filter TestRunFilter) *url.URL {
+func (a appEngineAPIImpl) GetRunsURL(filter TestRunFilter) *url.URL {
 	return getURL(a.GetHostname(), "/runs", filter)
 }
 
-// GetResultsUploadURL returns a url for uploading results to wpt.fyi.
-func (a AppEngineAPIImpl) GetResultsUploadURL() *url.URL {
+func (a appEngineAPIImpl) GetResultsUploadURL() *url.URL {
 	result, _ := url.Parse(fmt.Sprintf("https://%s%s", a.GetVersionedHostname(), "/api/results/upload"))
 	return result
 }
 
-// GetResultsURL returns a url for the wpt.fyi results page for the test runs
-// loaded for the given filter.
 func getURL(host, path string, filter TestRunFilter) *url.URL {
 	detailsURL, _ := url.Parse(fmt.Sprintf("https://%s%s", host, path))
 	detailsURL.RawQuery = filter.ToQuery().Encode()
