+++ webapp/routes.go
@@ -11,6 +11,7 @@ import (
 	"github.com/web-platform-tests/wpt.fyi/api/azure"
 	"github.com/web-platform-tests/wpt.fyi/api/checks"
 	"github.com/web-platform-tests/wpt.fyi/api/query"
+	"github.com/web-platform-tests/wpt.fyi/api/receiver"
 	"github.com/web-platform-tests/wpt.fyi/api/screenshot"
 	"github.com/web-platform-tests/wpt.fyi/api/taskcluster"
 	"github.com/web-platform-tests/wpt.fyi/shared"
@@ -24,6 +25,7 @@ func init() {
 	azure.RegisterRoutes()
 	checks.RegisterRoutes()
 	query.RegisterRoutes()
+	receiver.RegisterRoutes()
 	screenshot.RegisterRoutes()
 	taskcluster.RegisterRoutes()
 	RegisterRoutes()
